added creator cell bindings
This has brought things a long way, but has also introduced a new issue. Newly appended objects (with new events) need to be registered for each user. For example, if we add a cell on one client and the other client begins to type into it they will get an error -- because the events for `cellinput` (in `Olive's case, `cellhighlight`, were never handled)

As you can imagine, this took me quite a while to figure out, lol.