Forced Julia `1.11` for this specific extension
Something makes this incompatible with `1.8`, I am pretty sure, and that was the issue during registration. This was tested with `1.11.5`, so I figured I would bind compatibility high. If you want to use this package in your `Olive`, you will need a recent version of Julia (1.11 is currently a version behind)